requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
boto3==1.38.4
botocore==1.38.4
Brotli==1.1.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==45.0.4
cssselect2==0.8.0
daphne==4.2.0
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.4
django-admin-rangefilter==0.13.2
django-cors-headers==4.6.0
django-import-export==4.3.4
django-querycount==0.8.3
django-rest-swagger==2.2.0
django-storages==1.14.6
django-timezone-field==7.1
django-unfold==0.52.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
fonttools==4.56.0
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.47.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
msgpack==1.1.1
openapi-codec==1.3.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
s3transfer==0.12.0
sentry-sdk==2.22.0
service-identity==24.2.0
setuptools==80.9.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.0
webencodings==0.5.1
whitenoise==6.8.2
zope.interface==7.2
zopfli==0.2.3.post1
